PH-Regional-Inequality/requirements.txt
pandas>=2.1
numpy>=1.26
matplotlib>=3.8
seaborn>=0.13
plotly>=5.18
geopandas>=0.14
folium>=0.15
sqlalchemy>=2.0
psycopg2-binary>=2.9
jupyter>=1.0
nbconvert>=7.0
requests>=2.31
scipy>=1.11
openpyxl>=3.1
streamlit>=1.35
